fibo

+++ functions_nested_loops/102-fibonacci.c
@@ -7,16 +7,16 @@
 
 int main(void)
 {
-	int a = 1;
-	int b = 2;
-	int c;
+	long a = 1;
+	long b = 2;
+	long c;
 	int i;
 
 	printf("1, 2, "); 
 	for (i = 0; i < 49; i++)
 	{
 		c = a + b;
-		printf("%d", c);
+		printf("%ld", c);
 		if (i < 48)
 			printf(", ");
 		a = b;
